Bib: fix reference [30] (szegedy:2014:intriguing) with page range (commented)
We think it is best to not put any page range here, for the
following reason. The ICLR conference, where this article was published, has
online-only proceedings (https://iclr.cc/archive/2014/conference-proceedings/).
On that page, they link to the respective arXiv preprint of each article. The
one in question is 1312.6199.
However, if IEEE really requires adding a page range for this entry, then the
following may be used: 1--10.